requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-cloud-firestore